fix NumPy warnings on null vector instantiation (NaN as value for int array); updates to match newer Numba deprecations; new simpler clock() impl; NUMBA_OPT for CI; Python 3.11 on CI (#416)

+++ tests/unit_tests/test_traversals_with_bc_periodic.py
@@ -68,9 +68,9 @@ class TestPeriodicBoundaryCondition:
     @pytest.mark.parametrize(
         "data",
         (
-            np.array([1, 2, 3]),
-            np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]]),
-            np.arange(3 * 4 * 5).reshape((3, 4, 5)),
+            np.array([1.0, 2, 3]),
+            np.array([[1.0, 2, 3], [4, 5, 6], [7, 8, 9]]),
+            np.arange(3 * 4 * 5, dtype=float).reshape((3, 4, 5)),
         ),
     )
     @pytest.mark.parametrize("halo", (1, 2, 3))
@@ -123,15 +123,20 @@ class TestPeriodicBoundaryCondition:
     @pytest.mark.parametrize(
         "data",
         (
-            (np.array([1, 2, 3]),),
+            (np.array([1, 2, 3], dtype=float),),
             (
-                np.array([[41, 42, 33], [51, 52, 53], [61, 62, 63], [71, 72, 73]]),
-                np.array([[11, 12, 13, 14], [21, 22, 23, 24], [31, 32, 33, 34]]),
+                np.array(
+                    [[41, 42, 33], [51, 52, 53], [61, 62, 63], [71, 72, 73]],
+                    dtype=float,
+                ),
+                np.array(
+                    [[11, 12, 13, 14], [21, 22, 23, 24], [31, 32, 33, 34]], dtype=float
+                ),
             ),
             (
-                np.arange(4 * 4 * 5).reshape((3 + 1, 4, 5)),
-                np.arange(3 * 5 * 5).reshape((3, 4 + 1, 5)),
-                np.arange(3 * 4 * 6).reshape((3, 4, 5 + 1)),
+                np.arange(4 * 4 * 5, dtype=float).reshape((3 + 1, 4, 5)),
+                np.arange(3 * 5 * 5, dtype=float).reshape((3, 4 + 1, 5)),
+                np.arange(3 * 4 * 6, dtype=float).reshape((3, 4, 5 + 1)),
             ),
         ),
     )
